+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/listbox/DefaultListBoxBuilder.java
@@ -66,8 +66,9 @@ final class DefaultListBoxBuilder<T>
 
 	@Override
 	protected JComboBox<T> createComponent() {
-		FilterComboBoxModel<T> comboBoxModel = FilterComboBoxModel.<T>builder().build();
-		linkedValue.get().forEach(comboBoxModel.items()::add);
+		FilterComboBoxModel<T> comboBoxModel = FilterComboBoxModel.builder()
+						.items(linkedValue.get())
+						.build();
 		ListComboBox<T> comboBox = new ListComboBox<>(comboBoxModel, itemValue, linkedValue);
 		comboBox.setEditor(new Editor<>(itemValue));
 		comboBox.setEditable(true);